Fixed errors due to commit #119

+++ msdk-datamodel/src/main/java/io/github/msdk/datamodel/featuretables/ColumnName.java
@@ -38,7 +38,7 @@ public enum ColumnName {
             RTEND("RT End", Double.class), //
             DURATION("Duration", Double.class), //
             AREA("Area", Double.class), //
-            HEIGHT("Height", Double.class), //
+            HEIGHT("Height", Float.class), //
             CHARGE("Charge", Integer.class), //
             NUMBEROFDATAPOINTS("# Data Points", Integer.class), //
             FWHM("FWHM", Double.class), //
